runtime/wired_control/requirements.txt
pygame
pyserial
numpy
